Get video processing details

Retrieves the processing details for a vector of video IDs

Usage

get_video_processing_details(
  video_ids,
  client = yt_construct_client(),
  cache_disk = getOption("yt_cache_disk", FALSE),
  cache_key = getOption("yt_cache_key", NULL),
  token = NULL
)

Arguments

video_ids

A character vector of video ids.

client

A YouTube OAuth client created with yt_construct_client().

cache_disk

Should the access token be cached on disk? This reduces the number of times that you need to re-authenticate at the cost of storing access credentials on disk.

Learn more in https://httr2.r-lib.org/articles/oauth.html.

cache_key

If you are authenticating with multiple users using the same client, use this key to differentiate between those users.

token

A YouTube API OAuth token, or the access_token string from such a token. We recommend that you instead supply a client, in which case an appropriate token will be located if possible.

Value

A list of processing status details. processingStatus is probably of most interest

Examples

if (FALSE) { # yt_has_client_envvars() && interactive()
get_upload_playlist_id() |>
  get_playlist_items() |>
  get_video_processing_details()
}