Video snippet

Basic details about a video, including title, description, uploader, thumbnails and category.

Usage

yt_schema_video_snippet(
  title = character(),
  description = character(),
  tags = character(),
  category_id = character(),
  default_language = character()
)

Arguments

title

(character scalar) The video's title.

description

(character scalar) The video's description.

tags

(character vector) A list of keyword tags associated with the video. Tags may contain spaces.

category_id

(character scalar) The YouTube video category associated with the video.

default_language

(character scalar) The language of the videos's default snippet.

Value

A list with any non-zero-length properties.

Examples

yt_schema_video_snippet()
#> named list()
yt_schema_video_snippet(title = "An example video", tags = c("a", "b"))
#> $title
#> [1] "An example video"
#> 
#> $tags
#> [1] "a" "b"
#>