Video status

The status object contains information about the video's uploading, processing, and privacy statuses.

Usage

yt_schema_video_status(
  embeddable = logical(),
  license = character(),
  privacy_status = character(),
  public_stats_viewable = logical(),
  publish_at = datetime(),
  self_declared_made_for_kids = logical()
)

Arguments

embeddable

(logical) This value indicates if the video can be embedded on another website.

license

(factor scalar) The video's license. values: youtube, creativeCommon

privacy_status

(factor scalar) The video's privacy status. values: public, unlisted, private

public_stats_viewable

(logical) This value indicates if the extended video statistics on the watch page can be viewed by everyone. Note that the view count, likes, etc will still be visible if this is disabled.

publish_at

(datetime scalar) The date and time when the video is scheduled to publish. It can be set only if the privacy status of the video is private.

self_declared_made_for_kids

(logical) Whether the video is made for kids.

Value

A list with any non-zero-length properties.

Examples

yt_schema_video_status()
#> named list()
yt_schema_video_status(embeddable = TRUE)
#> $embeddable
#> [1] TRUE
#>